+++ tests/__init__.py
@@ -36,8 +36,9 @@ def suite(*args, **kw):
     from tests import test_specplot
     from tests import test_specplot_gallery
     from tests import test_utils
+    from tests import test_uxml
     from tests import test_writer
-    from tests import test_XPCSplugin
+    from tests import test_XPCS
     from tests import issue64
     from tests import issue99_hklscan
     from tests import issue107
@@ -58,8 +59,9 @@ def suite(*args, **kw):
         test_specplot,
         test_specplot_gallery,
         test_utils,
+        test_uxml,
         test_writer,
-        test_XPCSplugin,
+        test_XPCS,
         issue64,
         issue99_hklscan,
         issue107,
@@ -77,3 +79,4 @@ def suite(*args, **kw):
 if __name__ == '__main__':
     runner=unittest.TextTestRunner(verbosity=2)
     runner.run(suite())
+